+++ src/Silice3D/camera/third_personal_camera.hpp
@@ -10,16 +10,16 @@ namespace Silice3D {
 class ThirdPersonalCamera : public PerspectiveCamera {
  public:
   ThirdPersonalCamera(GameObject* parent,
-                      float fov,
-                      float z_near,
-                      float z_far,
-                      const glm::vec3& position,
-                      float mouse_sensitivity = 1.0,
-                      float mouse_scroll_sensitivity = 1.0,
-                      float min_dist_mod = 0.25,
-                      float max_dist_mod = 4.00,
-                      float base_distance = 0.0,
-                      float dist_offset = 0.0);
+                      double fov,
+                      double z_near,
+                      double z_far,
+                      const glm::dvec3& position,
+                      double mouse_sensitivity = 1.0,
+                      double mouse_scroll_sensitivity = 1.0,
+                      double min_dist_mod = 0.25,
+                      double max_dist_mod = 4.00,
+                      double base_distance = 0.0,
+                      double dist_offset = 0.0);
 
   virtual ~ThirdPersonalCamera() {}
 
@@ -31,12 +31,12 @@ class ThirdPersonalCamera : public PerspectiveCamera {
   bool first_call_;
 
   // Private constant numbers
-  const float initial_distance_, base_distance_, cos_max_pitch_angle_,
+  const double initial_distance_, base_distance_, cos_max_pitch_angle_,
                mouse_sensitivity_, mouse_scroll_sensitivity_,
                min_dist_mod_, max_dist_mod_, dist_offset_;
 
   // For mouseScrolled interpolation
-  float curr_dist_mod_, dest_dist_mod_;
+  double curr_dist_mod_, dest_dist_mod_;
 
   virtual void Update() override;
   virtual void MouseScrolled(double, double yoffset) override;
